refactor: template 이미지 목데이터 분리
